+++ test_package/test_package.cpp
@@ -3,6 +3,8 @@
 
 int main()
 {
-    const fdeep::tensor t(fdeep::tensor_shape(4), {1, 2, 3, 4});
+    const fdeep::tensor t(
+        fdeep::tensor_shape(static_cast<std::size_t>(4)),
+        {1, 2, 3, 4});
     std::cout << fdeep::show_tensor(t) << std::endl;
 }
